pyproject.toml
[tool.poetry]
name = "uptodown"
version = "0.1.0"
description = ""
authors = ["mandicoder <mandicoder@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
bs4 = "^0.0.2"
pyshorteners = "^1.0.1"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
wget = "^3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
